1_Ano/2_Semestre/EGS/Projeto/Auth/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
flasgger==0.9.5
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.2.2
kafka-python==2.0.2
psycopg2-binary
six==1.16.0
PyJWT==2.8.0
authlib
